+++ src/targets/jsr.ts
@@ -63,6 +63,5 @@ export const JSR: Target = <Target>{
 		next()
 	},
 	internalRules: internal,
-	isIgnoreFile: (path) => extractors.some((e) => e.path === path),
 	root: ".",
 }