+++ src/targets/vsce.ts
@@ -91,6 +91,5 @@ export const VSCE: Target = <Target>{
 		})
 	},
 	internalRules: internal,
-	isIgnoreFile: (path) => extractors.some((e) => e.path === path),
 	root: ".",
 }