+++ src/targets/yarn.ts
@@ -123,6 +123,5 @@ export const Yarn: Target = <Target>{
 		})
 	},
 	internalRules: internal,
-	isIgnoreFile: (path) => extractors.some((e) => e.path === path),
 	root: ".",
 }