+++ src/targets/yarnClassic.ts
@@ -113,6 +113,5 @@ export const YarnClassic: Target = <Target>{
 		})
 	},
 	internalRules: internal,
-	isIgnoreFile: (path) => extractors.some((e) => e.path === path),
 	root: ".",
 }