+++ src/testStream.test.ts
@@ -10,20 +10,24 @@ describe("Git", () => {
 			{ file: "1", src: { file: "2" } },
 			({ stream }) => {
 				const paths: { kind: number; path: string }[] = []
-				stream.addListener("dirent", (d) => {
+				stream.addEventListener("dirent", ({ detail: d }) => {
 					paths.push({ kind: d.match.kind, path: d.path })
 				})
-				stream.once("end", () => {
-					expect(paths.sort((a, b) => a.path.localeCompare(b.path))).toMatchObject(
-						[
-							{ kind: RuleMatchKind.missingSource, path: "file" },
-							{ kind: RuleMatchKind.missingSource, path: "src/" },
-							{ kind: RuleMatchKind.missingSource, path: "src/" },
-							{ kind: RuleMatchKind.missingSource, path: "src/file" },
-						].sort((a, b) => a.path.localeCompare(b.path)),
-					)
-					done()
-				})
+				stream.addEventListener(
+					"end",
+					() => {
+						expect(paths.sort((a, b) => a.path.localeCompare(b.path))).toMatchObject(
+							[
+								{ kind: RuleMatchKind.missingSource, path: "file" },
+								{ kind: RuleMatchKind.missingSource, path: "src/" },
+								{ kind: RuleMatchKind.missingSource, path: "src/" },
+								{ kind: RuleMatchKind.missingSource, path: "src/file" },
+							].sort((a, b) => a.path.localeCompare(b.path)),
+						)
+						done()
+					},
+					{ once: true },
+				)
 			},
 			{ target },
 		)
@@ -33,22 +37,26 @@ describe("Git", () => {
 			{ ".git": { HEAD: "" }, ".gitignore": "file", file: "1", src: { file: "2" } },
 			({ stream }) => {
 				const paths: { kind: number; path: string }[] = []
-				stream.addListener("dirent", (d) => {
+				stream.addEventListener("dirent", ({ detail: d }) => {
 					paths.push({ kind: d.match.kind, path: d.path })
 				})
-				stream.once("end", () => {
-					expect(paths.sort((a, b) => a.path.localeCompare(b.path))).toMatchObject(
-						[
-							{ kind: RuleMatchKind.external, path: "file" },
-							{ kind: RuleMatchKind.noMatch, path: "src/" },
-							{ kind: RuleMatchKind.external, path: "src/file" },
-							{ kind: RuleMatchKind.noMatch, path: ".gitignore" },
-							{ kind: RuleMatchKind.internal, path: ".git/" },
-							{ kind: RuleMatchKind.internal, path: ".git/HEAD" },
-						].sort((a, b) => a.path.localeCompare(b.path)),
-					)
-					done()
-				})
+				stream.addEventListener(
+					"end",
+					() => {
+						expect(paths.sort((a, b) => a.path.localeCompare(b.path))).toMatchObject(
+							[
+								{ kind: RuleMatchKind.external, path: "file" },
+								{ kind: RuleMatchKind.noMatch, path: "src/" },
+								{ kind: RuleMatchKind.external, path: "src/file" },
+								{ kind: RuleMatchKind.noMatch, path: ".gitignore" },
+								{ kind: RuleMatchKind.internal, path: ".git/" },
+								{ kind: RuleMatchKind.internal, path: ".git/HEAD" },
+							].sort((a, b) => a.path.localeCompare(b.path)),
+						)
+						done()
+					},
+					{ once: true },
+				)
 			},
 			{ target },
 		)